Arquitectura de capas (N-Tier) implementada
