--- back/src/main/java/com/app/configuration/CorsConfiguration.java
@@ -1,18 +1,0 @@
-package com.app.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CorsConfiguration implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("https://kzmk6so15813h5xcbdjk.lite.vusercontent.net:4200",
-                        "https://c24-61-n-webapp.onrender.com",
-                        "http://localhost:4200")
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS", "HEAD", "TRACE", "CONNECT");
-    }
-}
