--- back/src/main/java/com/app/domain/application/Application.java
@@ -1,46 +1,0 @@
-package com.app.domain.application;
-
-import com.app.domain.candidate.Candidate;
-import com.app.infra.enums.Status;
-import com.app.domain.job.Job;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "applications")
-public class Application {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long applicationId;
-    private long candidateId;
-    private long jobId;
-    private Status status;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-
-    @ManyToOne
-    private Candidate candidate;
-
-    @ManyToOne
-    private Job job;
-
-    @PrePersist
-    public void prePersist() {
-        this.createdAt = LocalDateTime.now();
-        this.updatedAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    public void preUpdate() {
-        this.updatedAt = LocalDateTime.now();
-    }
-}
