--- back/src/main/java/com/app/domain/candidate/CandidateRepository.java
@@ -1,6 +1,0 @@
-package com.app.domain.candidate;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CandidateRepository extends JpaRepository<Candidate, Long> {
-}
