--- back/src/main/java/com/app/domain/candidate/CandidateUpdater.java
@@ -1,21 +1,0 @@
-package com.app.domain.candidate;
-
-import com.app.domain.candidate.dto.CandidateUpdateDto;
-
-public class CandidateUpdater {
-    public static Candidate updateCandidateFields(Candidate candidate, CandidateUpdateDto payload) {
-        if (payload.name() != null) candidate.setName(payload.name());
-        if (payload.identification() != null) candidate.setIdentification(payload.identification());
-        if (payload.phoneNumber() != null) candidate.setPhoneNumber(payload.phoneNumber());
-        if (payload.email() != null) candidate.setEmail(payload.email());
-        if (payload.password() != null) candidate.setPassword(payload.password());
-        if (payload.position() != null) candidate.setPosition(payload.position());
-        if (payload.resume() != null) candidate.setResume(payload.resume());
-        if (payload.experienceLevel() != null) candidate.setExperienceLevel(payload.experienceLevel());
-        if (payload.skills() != null) candidate.setSkills(payload.skills());
-        if (payload.education() != null) candidate.setEducation(payload.education());
-        if (payload.location() != null) candidate.setLocation(payload.location());
-        if (payload.availableToTravel() != null) candidate.setAvailableToTravel(payload.availableToTravel());
-        return candidate;
-    }
-}
