+++ back/src/main/java/com/app/domain/dto/recruiter/RecruiterRegisterDto.java
@@ -1,0 +1,18 @@
+package com.app.domain.dto.recruiter;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record RecruiterRegisterDto(
+        @NotBlank
+        String name,
+        String identification,
+        String phoneNumber,
+        @NotBlank @Email
+        String email,
+        @NotBlank
+        String password,
+        String position,
+        Long companyId
+) {
+}
