--- back/src/main/java/com/app/domain/job/Job.java
@@ -1,56 +1,0 @@
-package com.app.domain.job;
-
-import com.app.domain.application.Application;
-import com.app.domain.company.Company;
-import com.app.infra.enums.EmploymentType;
-import com.app.domain.recruiter.Recruiter;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "jobs")
-public class Job {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long jobId;
-
-    private String title;
-    private String description;
-    private String requirements;
-    private float salary;
-    private EmploymentType employmentType;
-    private long companyId;
-    private long recruiterId;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-
-    @ManyToOne
-    private Recruiter recruiter;
-
-    @OneToMany(mappedBy = "job", cascade = CascadeType.ALL)
-    private List<Application> applications;
-
-    @ManyToOne
-    private Company company;
-
-    @PrePersist
-    public void prePersist() {
-        this.createdAt = LocalDateTime.now();
-        this.updatedAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    public void preUpdate() {
-        this.updatedAt = LocalDateTime.now();
-    }
-}
