+++ back/src/main/java/com/app/domain/model/Location.java
@@ -1,0 +1,5 @@
+package com.app.domain.model;
+
+public enum Location {
+    MADRID,BARCELONA,VALENCIA,REMOTE
+}
