+++ back/src/main/java/com/app/domain/model/Recruiter.java
@@ -1,0 +1,27 @@
+package com.app.domain.model;
+
+import com.app.domain.model.job.Job;
+import com.app.domain.model.user.User;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@DiscriminatorValue("RECRUITER")
+public class Recruiter extends User {
+    private long companyId;
+
+    @OneToMany(mappedBy = "recruiter", cascade = CascadeType.ALL)
+    private List<Job> jobs;
+
+    @ManyToOne
+    private Company company;
+}
