+++ back/src/main/java/com/app/domain/model/application/Application.java
@@ -1,0 +1,45 @@
+package com.app.domain.model.application;
+
+import com.app.domain.model.Candidate;
+import com.app.domain.model.job.Job;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "applications")
+public class Application {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long applicationId;
+    private long candidateId;
+    private long jobId;
+    private Status status;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+
+    @ManyToOne
+    private Candidate candidate;
+
+    @ManyToOne
+    private Job job;
+
+    @PrePersist
+    public void prePersist() {
+        this.createdAt = LocalDateTime.now();
+        this.updatedAt = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    public void preUpdate() {
+        this.updatedAt = LocalDateTime.now();
+    }
+}
