+++ back/src/main/java/com/app/domain/model/application/Status.java
@@ -1,0 +1,5 @@
+package com.app.domain.model.application;
+
+public enum Status {
+    PENDING,INTERVIEW,REJECTED,HIRED
+}
