+++ back/src/main/java/com/app/domain/model/job/EmploymentType.java
@@ -1,0 +1,5 @@
+package com.app.domain.model.job;
+
+public enum EmploymentType {
+    REMOTE,FACE_TO_FACE,HYBRID
+}
