+++ back/src/main/java/com/app/domain/model/user/Role.java
@@ -1,0 +1,5 @@
+package com.app.domain.model.user;
+
+public enum Role {
+    CANDIDATE,RECRUITER,ADMIN
+}
