--- back/src/main/java/com/app/domain/recruiter/RecruiterRepository.java
@@ -1,6 +1,0 @@
-package com.app.domain.recruiter;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RecruiterRepository extends JpaRepository<Recruiter, Long> {
-}
