--- back/src/main/java/com/app/domain/recruiter/RecruiterService.java
@@ -1,10 +1,0 @@
-package com.app.domain.recruiter;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class RecruiterService {
-    private final RecruiterRepository recruiterRepository;
-}
