--- back/src/main/java/com/app/domain/user/UserRepository.java
@@ -1,8 +1,0 @@
-package com.app.domain.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    UserDetails findByEmail(String email);
-}
