--- back/src/main/java/com/app/infra/enums/EmploymentType.java
@@ -1,5 +1,0 @@
-package com.app.infra.enums;
-
-public enum EmploymentType {
-    REMOTE,FACE_TO_FACE,HYBRID
-}
