--- back/src/main/java/com/app/infra/enums/Location.java
@@ -1,5 +1,0 @@
-package com.app.infra.enums;
-
-public enum Location {
-    MADRID,BARCELONA,VALENCIA,REMOTE
-}
