--- back/src/main/java/com/app/infra/enums/Role.java
@@ -1,5 +1,0 @@
-package com.app.infra.enums;
-
-public enum Role {
-    CANDIDATE,RECRUITER,ADMIN
-}
