--- back/src/main/java/com/app/infra/enums/Status.java
@@ -1,5 +1,0 @@
-package com.app.infra.enums;
-
-public enum Status {
-    PENDING,INTERVIEW,REJECTED,HIRED
-}
