--- back/src/main/java/com/app/infra/security/SecurityFilter.java
@@ -1,44 +1,0 @@
-package com.app.infra.security;
-
-import com.app.domain.user.UserRepository;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@Component
-public class SecurityFilter extends OncePerRequestFilter {
-
-    @Autowired
-    private TokenService tokenService;
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException {
-
-        var jwt = request.getHeader("Authorization");
-
-        if (jwt != null) {
-            jwt = jwt.replace("Bearer ", "");
-            var subject = tokenService.getSubject(jwt);
-            if (subject != null) {
-                // Valid token
-                var user = userRepository.findByEmail(subject);
-                var authentication = new UsernamePasswordAuthenticationToken(user, null,
-                        user.getAuthorities());
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-            }
-        }
-
-        filterChain.doFilter(request, response);
-    }
-}
