+++ back/src/main/java/com/app/repository/CandidateRepository.java
@@ -1,0 +1,7 @@
+package com.app.repository;
+
+import com.app.domain.model.Candidate;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CandidateRepository extends JpaRepository<Candidate, Long> {
+}
