+++ back/src/main/java/com/app/repository/RecruiterRepository.java
@@ -1,0 +1,7 @@
+package com.app.repository;
+
+import com.app.domain.model.Recruiter;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RecruiterRepository extends JpaRepository<Recruiter, Long> {
+}
