+++ back/src/main/java/com/app/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.app.repository;
+
+import com.app.domain.model.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    UserDetails findByEmail(String email);
+}
