+++ back/src/main/java/com/app/service/RecruiterService.java
@@ -1,0 +1,11 @@
+package com.app.service;
+
+import com.app.repository.RecruiterRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class RecruiterService {
+    private final RecruiterRepository recruiterRepository;
+}
